requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
dataclasses-json==0.6.7
dnspython==2.6.1
einops==0.8.0
fastapi==0.115.0
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langsmith==0.1.126
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
marshmallow==3.22.0
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
motor==3.6.0
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict==6.1.0
mypy-extensions==1.0.0
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow @ file:///croot/pillow_1721059439630/work
psutil==6.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pymongo==4.9.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
sympy @ file:///croot/sympy_1724938189289/work
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.6
websockets==13.1
wheel==0.44.0
wikipedia==1.4.0
yarl==1.12.1
